Some compiler warnings removed.
git-svn-id: https://svn.aros.org/svn/aros/trunk@33419 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/prefs/Zune/zunestuff.c
@@ -45,7 +45,7 @@ long aslfilerequest(char *msg,char *dirpart,char *filepart,char *fullname, struc
         }; 
   
     
-        if ( (int)(fr = (struct FileRequester *) AllocAslRequest(ASL_FileRequest, frtags) ) )
+        if ( (fr = (struct FileRequester *) AllocAslRequest(ASL_FileRequest, frtags) ) )
         {
             if (AslRequest(fr, NULL))
             {
